requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
colorama==0.4.6
Django==5.2.11
django-redis==6.0.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
redis==7.1.1
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
